EskResult allows a single error to be passed to constructors, added error helper fn
